src/requirements.txt
apache-airflow==2.2.5
requests
# apache-airflow-providers-google
google-cloud-storage
google-cloud
notebook
opencv-python-headless
dvc[gs]
mlflow
# protobuf==3.19
numpy==1.23.5
optree==0.10.0
tensorflow
keras-tuner
pillow
scikit-image
scikit-learn
python-dotenv